pyproject.toml
[project]
name = "podcast-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "elevenlabs>=2.31.0",
    "google-genai>=1.60.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "rich>=14.3.1",
    "textual>=7.3.0",
]
